pyproject.toml
[project]
name = "tc-ml-books"
version = "0.1.0"
description = "Projeto para recomendações de livros."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "beautifulsoup4>=4.13.4",
    "fastapi~=0.116.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]~=0.35.0",
]
